requirements.txt
# AI Image Detection - Requirements
# Python 3.8+

# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0

# Image Processing
Pillow>=9.0.0

# Face Detection & Recognition
facenet-pytorch>=2.5.2
mtcnn>=0.1.1
opencv-python>=4.8.0

# Data Science & Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.0.0

# Utilities
tqdm>=4.65.0

# Web Server
Flask>=2.3.0

# Optional: For GUI (uncomment if needed)
# streamlit>=1.20.0
# gradio>=3.20.0
